### 2017-05-31 #116 @mpadge
bikedata
Download and aggregate data from all public hire bicycle systems which provide open data, currently including Santander Cycles in London, U.K., and from the U.S.A., citibike in New York City NY, Divvy in Chicago IL, Capital Bikeshare in Washington DC, Hubway in Boston MA, and Metro in Los Angeles LA.

```
Package: bikedata
Title: Download and Aggregate Data from Public Hire Bicycle Systems
Version: 0.0.1
    person("Mark", "Padgham", email = "mark.padgham@email.com", role = c("aut", "cre")),
    person("Richard", "Ellison", role = "aut"),
    person(family = "SQLite Consortium", role = "ctb",
        comment = "Authors of included SQLite code"))
Description: Download and aggregate data from all public hire bicycle systems
    which provide open data, currently including Santander Cycles in London,
    U.K., and from the U.S.A., citibike in New York City NY, Divvy in Chicago
    IL, Capital Bikeshare in Washington DC, Hubway in Boston MA, and Metro in
    Los Angeles LA.
    httr,
    Rcpp,
    RSQLite,
    reshape2,
    xml2
    roxygen2,
    covr
    BH,
    Rcpp
SystemRequirements: C++11
NeedsCompilation: yes
Remotes: jimhester/covr
```

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under and why? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):

Data retrieval

All those interested in analysing urban mobility - urban and general transport modellers and planners, urban scientists. Anyone wanting to make cool data visualisations of movement through some of the euro-american world's biggest cities (like [this prominent example](http://toddwschneider.com/posts/a-tale-of-twenty-two-million-citi-bikes-analyzing-the-nyc-bike-share-system/)).

No current R packages enable importing of public hire bicycle data. @ramnathv has his [visualising bike share repo](https://github.com/ramnathv/bikeshare), but that's not an R package. I've heard rumours of a package in development for live station feeds [along these lines](https://www.citybik.es/), but even if and when that appears, it would focus on real-time data (primarily bicycle availability), whereas `bikedata` focusses exclusively on archiving and aggregating historical data.

Package will be archived as soon as the ropensci review process has been completed, in accordance with JOSS policies.

@ramnathv because of his prominent prior work on analysing public hire bike systems

-----

Note that `goodpractice` flags only:
1. Test coverage, which will be extended; and
2. Large installed size, for explanation of which see [`cran-comments.md`](https://github.com/mpadge/bikedata/blob/master/cran-comments.md)
