### 2017-05-31 #116 @maelle
@mpadge short question even before the real editor checks, why do you use the development version of `covr`?
