### 2017-05-31 #116 @mpadge
That's not part of the CRAN submission (obviously), and is used because of Jim Hester's help [here](https://github.com/mpadge/bikedata/commit/12533f917bd617c1c00c27022a49cc3a0e7033bd) - see his extended commit message at top. It's temporary and will be resolved once `covr` is updated on CRAN.
