+++ src/utils.h
@@ -188,10 +188,12 @@ std::string convert_datetime_ch (std::string str)
     return str;
 }
 
-//' convert_datetime_la
+//' convert_datetime_la_ph
+//'
+//' LA and Philadelphia have identical formats
 //'
 //' @noRd
-std::string convert_datetime_la (std::string str)
+std::string convert_datetime_la_ph (std::string str)
 {
     std::string mon = str_token (&str, "/");
     if (mon.length () == 1)
