+++ R/RcppExports.R
@@ -5,7 +5,7 @@
 #'
 #' Inserts data into the table of stations in the database. Applies to those
 #' cities for which station data are included and read as part of the actual
-#' raw trips data: ny, boston, la.
+#' raw trips data: ny, boston, la, philly.
 #'
 #' @param dbcon Active connection to sqlite3 database
 #' @param stationqry Station query constructed during reading of data with
@@ -59,8 +59,7 @@ NULL
 #'
 #' @noRd
 rcpp_import_stn_df <- function(bikedb, stn_data, city) {
-    .Call('bikedata_rcpp_import_stn_df', PACKAGE = 'bikedata',
-          bikedb, stn_data, city)
+    .Call('bikedata_rcpp_import_stn_df', PACKAGE = 'bikedata', bikedb, stn_data, city)
 }
 
 #' rcpp_import_to_trip_table
@@ -79,8 +78,7 @@ rcpp_import_stn_df <- function(bikedb, stn_data, city) {
 #'
 #' @noRd
 rcpp_import_to_trip_table <- function(bikedb, datafiles, city, quiet) {
-    .Call('bikedata_rcpp_import_to_trip_table', PACKAGE = 'bikedata',
-          bikedb, datafiles, city, quiet)
+    .Call('bikedata_rcpp_import_to_trip_table', PACKAGE = 'bikedata', bikedb, datafiles, city, quiet)
 }
 
 #' rcpp_import_to_file_table
@@ -97,8 +95,7 @@ rcpp_import_to_trip_table <- function(bikedb, datafiles, city, quiet) {
 #'
 #' @noRd
 rcpp_import_to_file_table <- function(bikedb, datafiles, city, nfiles) {
-    .Call('bikedata_rcpp_import_to_file_table', PACKAGE = 'bikedata',
-          bikedb, datafiles, city, nfiles)
+    .Call('bikedata_rcpp_import_to_file_table', PACKAGE = 'bikedata', bikedb, datafiles, city, nfiles)
 }
 
 #' rcpp_create_sqlite3_db
@@ -131,8 +128,7 @@ rcpp_create_sqlite3_db <- function(bikedb) {
 #'
 #' @noRd
 rcpp_create_db_indexes <- function(bikedb, tables, cols, reindex) {
-    .Call('bikedata_rcpp_create_db_indexes', PACKAGE = 'bikedata',
-          bikedb, tables, cols, reindex)
+    .Call('bikedata_rcpp_create_db_indexes', PACKAGE = 'bikedata', bikedb, tables, cols, reindex)
 }
 
 #' rcpp_create_city_index
@@ -148,8 +144,7 @@ rcpp_create_db_indexes <- function(bikedb, tables, cols, reindex) {
 #'
 #' @noRd
 rcpp_create_city_index <- function(bikedb, reindex) {
-    .Call('bikedata_rcpp_create_city_index', PACKAGE = 'bikedata',
-          bikedb, reindex)
+    .Call('bikedata_rcpp_create_city_index', PACKAGE = 'bikedata', bikedb, reindex)
 }
 
 #' strtokm
@@ -211,7 +206,9 @@ NULL
 #' @noRd
 NULL
 
-#' convert_datetime_la
+#' convert_datetime_la_ph
+#'
+#' LA and Philadelphia have identical formats
 #'
 #' @noRd
 NULL
