### 2017-05-31 31f7fe7
fix philly entry in bikedata-files.R
