### 2017-05-31 #116 @maelle
Thanks for your submission @mpadge! I'm currently looking for reviewers and also bothering you/potentially other people with an R CHECK error, see comments.

### Editor checks:

- [x ] **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
- [ x] **License:** The package has a CRAN or OSI accepted license
- [ x] **Repository:** The repository link resolves correctly

---

#### Editor comments

* I (Windows user) could install the package. The first time I tried building it I got a weird error about g++ not being available, then I tried building it again and it worked.

* `devtools::test` works fine.

* R CHECK gives me an installation error, see install.out:

```
* installing *source* package 'bikedata' ...
** libs

*** arch - i386
C:/RBUILD~1/3.3/mingw_32/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c RcppExports.cpp -o RcppExports.o
C:/RBUILD~1/3.3/mingw_32/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c sqlite3db-setup.cpp -o sqlite3db-setup.o
In file included from sqlite3db-setup.cpp:22:0:
sqlite3db-add-data.h: In function 'int rcpp_import_to_trip_table(const char*, Rcpp::CharacterVector, std::string, bool)':
sqlite3db-add-data.h:162:31: warning: suggest parentheses around comparison in operand of '|' [-Wparentheses]
                 else if (city == "la" | city == "ph")
                               ^
sqlite3db-setup.cpp: In function 'int rcpp_create_db_indexes(const char*, Rcpp::CharacterVector, Rcpp::CharacterVector, bool)':
sqlite3db-setup.cpp:120:11: warning: variable 'sqliteversion' set but not used [-Wunused-but-set-variable]
     char *sqliteversion = (char *)"0.1";
           ^
C:\RBUILD~1\3.3\mingw_32\bin\nm.exe: sqlite3/sqlite3.o: File format not recognized
C:/RBUILD~1/3.3/mingw_32/bin/g++ -shared -s -static-libgcc -o bikedata.dll tmp.def RcppExports.o sqlite3db-setup.o sqlite3/sqlite3.o -Ld:/Compiler/gcc-4.9.3/local330/lib/i386 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/PROGRA~1/R/R-34~1.0PA/bin/i386 -lR
sqlite3/sqlite3.o: file not recognized: File format not recognized
collect2.exe: error: ld returned 1 exit status
no DLL was created
ERROR: compilation failed for package 'bikedata'
* removing 'C:/Users/Maelle/AppData/Local/Temp/RtmpoJOdNK/bikedata.Rcheck/bikedata'
```

* I found a typo, "demographc" in bike_demographic_data.Rd:10

---

Reviewers:
Due date:
