### 2017-05-31 43a7bde
remove unused sqliteversion stmts from src/
