### 2017-05-31 #116 @mpadge
Thanks @maelle, the warnings are indeed my fault - sorry about that, and hopefully fixed already. I've no idea why `sqlite3/sqlite3.o` would not be recognised because your machine built it in the first place. First let's clarify that the warnings are gone, and hope that that fixes the `.o` issue...
