### 2017-06-01 #116 @maelle
Thanks @mpadge. Results from `devtools::install()` (`devtools::check()` to follow)

```r
> devtools::install()
Installing bikedata
Using GitHub PAT from envvar GITHUB_PAT
Installing 1 package: Rcpp
Installing package into ‘C:/Users/Maelle/Documents/R/win-library/3.4’
(as ‘lib’ is unspecified)
trying URL 'https://cran.rstudio.com/bin/windows/contrib/3.4/Rcpp_0.12.11.zip'
Content type 'application/zip' length 3380318 bytes (3.2 MB)
downloaded 3.2 MB

package ‘Rcpp’ successfully unpacked and MD5 sums checked

The downloaded binary packages are in
	C:\Users\Maelle\AppData\Local\Temp\Rtmpqa31H0\downloaded_packages
"C:/PROGRA~1/R/R-34~1.0PA/bin/x64/R" --no-site-file --no-environ --no-save  \
  --no-restore --quiet CMD INSTALL "C:/Users/Maelle/Documents/bikedata"  \
  --library="C:/Users/Maelle/Documents/R/win-library/3.4" --install-tests

* installing *source* package 'bikedata' ...
** libs

*** arch - i386
C:/RBUILD~1/3.3/mingw_32/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c RcppExports.cpp -o RcppExports.o
C:/RBUILD~1/3.3/mingw_32/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c sqlite3db-setup.cpp -o sqlite3db-setup.o
C:/RBUILD~1/3.3/mingw_32/bin/gcc  -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O3 -Wall  -std=gnu99 -mtune=core2 -c sqlite3/sqlite3.c -o sqlite3/sqlite3.o
sqlite3/sqlite3.c: In function 'sqlite3ThreadJoin':
sqlite3/sqlite3.c:27032:8: warning: variable 'bRc' set but not used [-Wunused-but-set-variable]
   BOOL bRc;
        ^
C:/RBUILD~1/3.3/mingw_32/bin/g++ -shared -s -static-libgcc -o bikedata.dll tmp.def RcppExports.o sqlite3db-setup.o sqlite3/sqlite3.o -Ld:/Compiler/gcc-4.9.3/local330/lib/i386 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/PROGRA~1/R/R-34~1.0PA/bin/i386 -lR
installing to C:/Users/Maelle/Documents/R/win-library/3.4/bikedata/libs/i386

*** arch - x64
C:/RBUILD~1/3.3/mingw_64/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c RcppExports.cpp -o RcppExports.o
C:/RBUILD~1/3.3/mingw_64/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c sqlite3db-setup.cpp -o sqlite3db-setup.o
C:/RBUILD~1/3.3/mingw_64/bin/gcc  -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -std=gnu99 -mtune=core2 -c sqlite3/sqlite3.c -o sqlite3/sqlite3.o
sqlite3/sqlite3.c: In function 'sqlite3ThreadJoin':
sqlite3/sqlite3.c:27032:8: warning: variable 'bRc' set but not used [-Wunused-but-set-variable]
   BOOL bRc;
        ^
C:/RBUILD~1/3.3/mingw_64/bin/g++ -shared -s -static-libgcc -o bikedata.dll tmp.def RcppExports.o sqlite3db-setup.o sqlite3/sqlite3.o -Ld:/Compiler/gcc-4.9.3/local330/lib/x64 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/PROGRA~1/R/R-34~1.0PA/bin/x64 -lR
installing to C:/Users/Maelle/Documents/R/win-library/3.4/bikedata/libs/x64
** R
** data
*** moving datasets to lazyload DB
** inst
** tests
** preparing package for lazy loading
** help
*** installing help indices
** building package indices
** installing vignettes
** testing if installed package can be loaded
*** arch - i386
*** arch - x64
* DONE (bikedata)
```
