### 2017-06-01 #116 @maelle
`devtools::check()` -- @mpadge feel free to open an issue in your package repo if it's easier for you? Not sure what'd be best.

```r
Updating bikedata documentation
Loading bikedata
Re-compiling bikedata
"C:/PROGRA~1/R/R-34~1.0PA/bin/x64/R" --no-site-file --no-environ --no-save  \
  --no-restore --quiet CMD INSTALL "C:\Users\Maelle\Documents\bikedata"  \
  --library="C:\Users\Maelle\AppData\Local\Temp\Rtmpqa31H0\devtools_install_1a1431735fe9"  \
  --no-R --no-data --no-help --no-demo --no-inst --no-docs --no-exec --no-multiarch  \
  --no-test-load --preclean

* installing *source* package 'bikedata' ...
** libs
C:/RBUILD~1/3.3/mingw_64/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c RcppExports.cpp -o RcppExports.o
C:/RBUILD~1/3.3/mingw_64/bin/g++  -std=gnu++11 -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c sqlite3db-setup.cpp -o sqlite3db-setup.o
C:/RBUILD~1/3.3/mingw_64/bin/gcc  -I"C:/PROGRA~1/R/R-34~1.0PA/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE -I"C:/Users/Maelle/Documents/R/win-library/3.4/BH/include" -I"C:/Users/Maelle/Documents/R/win-library/3.4/Rcpp/include"   -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -std=gnu99 -mtune=core2 -c sqlite3/sqlite3.c -o sqlite3/sqlite3.o
sqlite3/sqlite3.c: In function 'sqlite3ThreadJoin':
sqlite3/sqlite3.c:27032:8: warning: variable 'bRc' set but not used [-Wunused-but-set-variable]
   BOOL bRc;
        ^
C:/RBUILD~1/3.3/mingw_64/bin/g++ -shared -s -static-libgcc -o bikedata.dll tmp.def RcppExports.o sqlite3db-setup.o sqlite3/sqlite3.o -Ld:/Compiler/gcc-4.9.3/local330/lib/x64 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/PROGRA~1/R/R-34~1.0PA/bin/x64 -lR
installing to C:/Users/Maelle/AppData/Local/Temp/Rtmpqa31H0/devtools_install_1a1431735fe9/bikedata/libs/x64
* DONE (bikedata)
Data for London, U.K. powered by TfL Open Data:
  Contains OS data <U+24B8> Crown copyright and database rights 2016
Data for New York City provided and owned by:
  NYC Bike Share, LLC and Jersey City Bike Share, LLC ("Bikeshare")
  see https://www.citibikenyc.com/data-sharing-policy
Data for Washington DC (Captialbikeshare), Chiago (Divvybikes) and Boston (Hubway)
  provided and owned by Motivate International Inc.
  see https://www.capitalbikeshare.com/data-license-agreement
  and https://www.divvybikes.com/data-license-agreement
  and https://www.thehubway.com/data-license-agreement
Writing NAMESPACE
Writing bikedata.Rd
Writing bike_test_data.Rd
Writing bike_db_totals.Rd
Writing bike_latest_files.Rd
Writing bike_datelimits.Rd
Writing bike_summary_stats.Rd
Writing bike_daily_trips.Rd
Writing bike_demographic_data.Rd
Writing dl_bikedata.Rd
Writing bike_stations.Rd
Writing store_bikedata.Rd
Writing bike_rm_db.Rd
Writing bike_tripmat.Rd
Writing bike_write_test_data.Rd
Writing bike_rm_test_data.Rd
Setting env vars --------------------------------------------------------------------
CFLAGS  : -Wall -pedantic
CXXFLAGS: -Wall -pedantic
Building bikedata -------------------------------------------------------------------
"C:/PROGRA~1/R/R-34~1.0PA/bin/x64/R" --no-site-file --no-environ --no-save  \
  --no-restore --quiet CMD build "C:\Users\Maelle\Documents\bikedata"  \
  --no-resave-data --no-manual

* checking for file 'C:\Users\Maelle\Documents\bikedata/DESCRIPTION' ... OK
* preparing 'bikedata':
* cleaning src
* creating vignettes ... OK
* cleaning src
* checking for LF line-endings in source and make files
* checking for empty or unneeded directories
* looking to see if a 'data/datalist' file should be added
* building 'bikedata_0.0.2.tar.gz'
Warning: file 'bikedata/cleanup' did not have execute permissions: corrected

Setting env vars --------------------------------------------------------------------
_R_CHECK_CRAN_INCOMING_ : FALSE
_R_CHECK_FORCE_SUGGESTS_: FALSE
Checking bikedata -------------------------------------------------------------------
"C:/PROGRA~1/R/R-34~1.0PA/bin/x64/R" --no-site-file --no-environ --no-save  \
  --no-restore --quiet CMD check  \
  "C:\Users\Maelle\AppData\Local\Temp\Rtmpqa31H0/bikedata_0.0.2.tar.gz" --as-cran  \
  --timings --no-manual

* using log directory 'C:/Users/Maelle/AppData/Local/Temp/Rtmpqa31H0/bikedata.Rcheck'
* using R version 3.4.0 Patched (2017-05-10 r72669)
* using platform: x86_64-w64-mingw32 (64-bit)
* using session charset: ISO8859-1
* using options '--no-manual --as-cran'
* checking for file 'bikedata/DESCRIPTION' ... OK
* this is package 'bikedata' version '0.0.2'
* package encoding: UTF-8
* checking package namespace information ... OK
* checking package dependencies ... NOTE
Package suggested but not available for checking: 'covr'
* checking if this is a source package ... NOTE
Found the following apparent object files/libraries:
  src/sqlite3/sqlite3.o
Object files/libraries should not be included in a source package.
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... OK
* checking whether package 'bikedata' can be installed ...Warning: running command '"C:/PROGRA~1/R/R-34~1.0PA/bin/x64/Rcmd.exe" INSTALL -l "C:/Users/Maelle/AppData/Local/Temp/Rtmpqa31H0/bikedata.Rcheck" --no-html "C:\Users\Maelle\AppData\Local\Temp\RTMPQA~1\BIKEDA~1.RCH\00_PKG~1\bikedata"' had status 1
 ERROR
Installation failed.
See 'C:/Users/Maelle/AppData/Local/Temp/Rtmpqa31H0/bikedata.Rcheck/00install.out' for details.
* DONE

Status: 1 ERROR, 2 NOTEs
See
  'C:/Users/Maelle/AppData/Local/Temp/Rtmpqa31H0/bikedata.Rcheck/00check.log'
for details.

R CMD check results
1 error  | 0 warnings | 2 notes
checking whether package 'bikedata' can be installed ... ERROR
Installation failed.
See 'C:/Users/Maelle/AppData/Local/Temp/Rtmpqa31H0/bikedata.Rcheck/00install.out' for details.

checking package dependencies ... NOTE
Package suggested but not available for checking: 'covr'

checking if this is a source package ... NOTE
Found the following apparent object files/libraries:
  src/sqlite3/sqlite3.o
Object files/libraries should not be included in a source package.
```
