### 2017-06-02 mpadge/bikedata#38 @mpadge
add explicit function to add indexes
Indexes for both times and dates as well as demographic characteristics could then be explicitly constructed
