### 2017-06-02 68554a6
mv src/sqlite3/ -> src/vendor/sqlite3
This is the easiest way to get git programming language statistics to **ignore** the sqlite3 files,
as explained in issue#37
