### 2017-06-02 a197be1
add city names to bike_summary_stats
