+++ R/database-stats.R
@@ -297,10 +297,10 @@ bike_summary_stats <- function (bikedb)
         latest_files <- c (NA, latest)
     }
 
-    res <- data.frame (num_trips = num_trips, num_stations = num_stations,
+    res <- data.frame (city = as.character (c ('total', cities)),
+                       num_trips = num_trips, num_stations = num_stations,
                        first_trip = dates [, 1], last_trip = dates [, 2],
                        latest_files = latest_files)
-    rownames (res) <- rnames
     return (tibble::as_tibble (res))
 }
 
