### 2017-06-02 #116 @maelle
Other PC, similar error when checking, this is install.out:

```
* installing *source* package 'bikedata' ...
** libs

*** arch - i386
c:/Rtools/mingw_32/bin/g++  -std=c++0x -I"C:/Users/MSALMO~1.ISG/DOCUME~1/R/R-33~1.1/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE   -I"C:/Users/msalmon.ISGLOBAL/Documents/R/win-library/3.3/BH/include" -I"C:/Users/msalmon.ISGLOBAL/Documents/R/win-library/3.3/Rcpp/include" -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c RcppExports.cpp -o RcppExports.o
c:/Rtools/mingw_32/bin/g++  -std=c++0x -I"C:/Users/MSALMO~1.ISG/DOCUME~1/R/R-33~1.1/include" -DNDEBUG -I. -DRSQLITE_USE_BUNDLED_SQLITE   -I"C:/Users/msalmon.ISGLOBAL/Documents/R/win-library/3.3/BH/include" -I"C:/Users/msalmon.ISGLOBAL/Documents/R/win-library/3.3/Rcpp/include" -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c sqlite3db-setup.cpp -o sqlite3db-setup.o
C:\Rtools\mingw_32\bin\nm.exe: vendor/sqlite3/sqlite3.o: File format not recognized
c:/Rtools/mingw_32/bin/g++ -shared -s -static-libgcc -o bikedata.dll tmp.def RcppExports.o sqlite3db-setup.o vendor/sqlite3/sqlite3.o -Ld:/Compiler/gcc-4.9.3/local330/lib/i386 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/Users/MSALMO~1.ISG/DOCUME~1/R/R-33~1.1/bin/i386 -lR
vendor/sqlite3/sqlite3.o: file not recognized: File format not recognized
collect2.exe: error: ld returned 1 exit status
no DLL was created
ERROR: compilation failed for package 'bikedata'
* removing 'C:/Users/MSALMO~1.ISG/AppData/Local/Temp/RtmpkzvaRb/bikedata.Rcheck/bikedata'
```

:thinking:
