### 2017-06-02 #116 @maelle
One just needs to check with the `--no-multiarch` option. :relieved:
