### 2017-06-03 #116 @maelle
Thanks @eamcvey @chucheria for accepting to review this package! 😺  As a reminder here are [the review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md) and the [reviewing guidelines](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md).

@mpadge can you confirm that/when you've changed the cleanup process? (related to what @jeroen said "The cleanup script should usually only clean files generated by configure. Files generated by make should be cleaned via a clean section in Makevars" / or use cleanup.win as mentioned by Jim Hester)
