### 2017-06-03 #116 @mpadge
Thanks @maelle, and thanks in advance @eamcvey and @chucheria for agreeing to review - exciting to hear what you might think of it! I've done the Jim Hester `cleanup.win` option. Jeroen is nevertheless correct that `cleanup` should only relate to `configure`-generated files, and I will definitely transition to a single `Makevars` soon. The whole build is nevertheless clean from here on for all OS's.
