+++ vignettes/bikedata.Rmd
@@ -34,6 +34,7 @@ Chicago IL, U.S.A. | [Divvy](https://www.divvybikes.com/) | 5,837 | 576
 Washingon DC, U.S.A. | [Capital BikeShare](https://www.capitalbikeshare.com/) | 4,457 | 406
 Boston MA, U.S.A. | [Hubway](https://www.thehubway.com/) | 1,461 | 158
 Los Angeles CA, U.S.A. | [Metro](https://bikeshare.metro.net/) | 1,000 | 65
+Philadelphia PA, U.S.A. | [Indego](https://www.rideindego.com) | 600 | 60
 
 All of these systems record and disseminate individual trip data, minimally
 including the times and places at which every trip starts and ends. Some provide
