### 2017-06-06 4f04343
update travis.yml to fix tlmgr update err
