### 2017-06-06 e2a647e
tidy README.md
