### 2017-06-12 4b05d47
tools::package_native_routine_registration_skeleton
https://stackoverflow.com/questions/42313373/r-cmd-check-note-found-no-calls-to-r-registerroutines-r-usedynamicsymbols#42339658
