+++ R/bikedata-package.R
@@ -64,5 +64,5 @@
 #' @importFrom tibble as_tibble tibble
 #' @importFrom utils data menu read.csv tail type.convert unzip write.csv zip
 #' @importFrom xml2 xml_children xml_find_all
-#' @useDynLib bikedata
+#' @useDynLib bikedata, .registration = TRUE
 NULL
