+++ DESCRIPTION
@@ -25,6 +25,7 @@ Imports:
     tibble,
     xml2
 Suggests:
+    dbplyr,
     knitr,
     rmarkdown,
     roxygen2,
