+++ NAMESPACE
@@ -44,4 +44,4 @@ importFrom(utils,write.csv)
 importFrom(utils,zip)
 importFrom(xml2,xml_children)
 importFrom(xml2,xml_find_all)
-useDynLib(bikedata)
+useDynLib(bikedata, .registration = TRUE)
