### 2017-06-12 2019eb2
update travis.yml
