### 2017-06-14 71baedd
automate dl of ph, la (issue#35)
