### 2017-06-18 #116 @chucheria
4h

---

It is a great package overall, it removes the necessity of have knowledge of many bike data APIs/file system of open data, which is a great time and effort saver. It will make more sense as more bike open data is available inside the package.

Vignette has example uses and is very complete.

I played with the functions a bit, it is a simple package and it allows you to use API information locally.

#### General code comments

I found some errors executing the code.

I downloaded NY data from 2016-01 to 2016-05 once I changed to Los Angeles city, it told me that all data already exist.
```r
dl_bikedata (city = 'New York City USA', dates = 201601:201605)
    Downloading 201605-citibike-tripdata.zip
    Downloading JC-201601-citibike-tripdata.csv.zip
    Downloading JC-201602-citibike-tripdata.csv.zip
    Downloading JC-201603-citibike-tripdata.csv.zip
    Downloading JC-201604-citibike-tripdata.csv.zip
    Downloading JC-201605-citibike-tripdata.csv.zip

dl_bikedata(city = 'la', dates = 201601:201605)
    All data files already exist
```

Once I downloaded NY data, I wanted to store it with `store_bikedata`:
```r
ny <- store_bikedata (city = 'nyc', bikedb = 'bikedb', dates = 201601:201605)
    Downloading data for nyc
    All data files already exist
    Adding data to sqlite3 database
    Unzipping raw data files ...
     Show Traceback

     Rerun with Debug
     Error in unzip(f, list = TRUE) :
      zip file '/var/folders/zc/yccgnjm528j4x_665t_rz0km0000gn/T//RtmpU6AJg6/201603-citibike-tripdata.zip' cannot be opened
```

If I do the `store_bikedata` from the start everything runs great.

#### Installation:

**macOS Sierra**:

Installation in macOS was smooth. A warning popped up in installation and in every test:
`Warning: Installed Rcpp (0.12.11) different from Rcpp used to build dplyr (0.12.10).
Please reinstall dplyr to avoid random crashes or undefined behavior.` I reinstalled `dplyr` and the warning dissapeared.

#### Documentation:

- There's no CONTRIBUTION file nor community guidelines to contribute, consider writing one to welcome people to contribute.
- There's no CODE_OF_CONDUCT file.
- There's no NEWS file with latest changes.
- LICENSE is also missing, which is also important, especially for you!
- 👍 All functions are explained with examples in the documentation. Documentation and the vignette look good.
- 👍 I ran the tests and everything looks like it is in order.
- 👍 Messages and warnings look good.

#### Scripts

Scripts are named with a hyphen instead of snake_case, something that rOpenSci recommends and packages seem to follow that pattern. Functions and variables are well named and explained.
