### 2017-06-18 #116 @chucheria
Please, tell me if I didn't explained well any of the comments or errors.
