### 2017-06-22 6a6412e
use RSQLite::dbConnect instead of dplyr
