### 2017-07-02 b201ebf
drop dplyr dep
dplyr 0.7.0 broke package, so it's been dropped in favour of clearer and more direct use of
RSQLite functions.
