+++ R/database-stats.R
@@ -62,7 +62,7 @@ bike_cities_in_db <- function (bikedb)
 get_new_datafiles <- function (bikedb, flist_zip)
 {
     db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
-    old_files <- dplyr::collect (dplyr::tbl (db, 'datafiles'))$name
+    old_files <- RSQLite::dbReadTable (db, 'datafiles')$name
     RSQLite::dbDisconnect (db)
     flist_zip [which (!basename (flist_zip) %in% old_files)]
 }
