+++ NEWS.md
@@ -1,0 +1,9 @@
+
+0.0.2
+===================
+- No change to package functionality
+- Drop dplyr dependency after dplyr 0.7 upgrade
+
+0.0.1 (31 May 2017)
+===================
+- Initial CRAN release
