+++ DESCRIPTION
@@ -15,9 +15,9 @@ License: GPL-3
 Depends:
     R (>= 3.0)
 Imports:
-    dplyr,
     httr,
     lubridate,
+    magrittr,
     methods,
     Rcpp,
     RSQLite,
