+++ NAMESPACE
@@ -21,15 +21,12 @@ importFrom(RSQLite,dbDisconnect)
 importFrom(RSQLite,dbGetQuery)
 importFrom(RSQLite,dbSendQuery)
 importFrom(Rcpp,evalCpp)
-importFrom(dplyr,"%>%")
-importFrom(dplyr,collect)
-importFrom(dplyr,src_sqlite)
-importFrom(dplyr,tbl)
 importFrom(httr,GET)
 importFrom(httr,content)
 importFrom(lubridate,ddays)
 importFrom(lubridate,interval)
 importFrom(lubridate,ymd)
+importFrom(magrittr,"%>%")
 importFrom(methods,as)
 importFrom(reshape2,dcast)
 importFrom(tibble,as_tibble)
