### 2017-07-06 ropensci/bikedata#42 @mpadge
call-a-bike data?
Thanks @kruse-alex for the heads up. These data are different to most, and seem restricted to Hamburg only - right Alex? But it reads as if they intend to maintain ongoing releases on an annual basis, making them acceptable for incorporation in `bikedata`. Alex: I've put this issue here in my package, becuase all of the infrastructure for reading and storing is there, so it'll be the easiest way for us both to get the data into `R`. Please jump in and help!  Note also that the repo should soon move to [`ropensci`](https://github.com/ropensci/onboarding/issues/116) (pending review).
