+++ NEWS.md
@@ -1,3 +1,5 @@
+0.0.3
+===================
 
 0.0.2
 ===================
