### 2017-07-13 2018fdd
ensure all db var types are consistent
and also fix LA data tests for historically defined period now that they've finally updated
