+++ src/sqlite3db-setup.cpp
@@ -57,20 +57,20 @@ int rcpp_create_sqlite3_db (const char * bikedb)
         "trip_duration numeric,"
         "start_time timestamp without time zone,"
         "stop_time timestamp without time zone,"
-        "start_station_id integer,"
-        "end_station_id integer,"
-        "bike_id integer,"
+        "start_station_id text,"
+        "end_station_id text,"
+        "bike_id text,"
         "user_type text,"
-        "birth_year integer,"
-        "gender integer"
+        "birth_year text,"
+        "gender text"
         ");"
         "CREATE TABLE stations ("
         "    id integer primary key,"
         "    city text,"
         "    stn_id text,"
         "    name text,"
-        "    longitude numeric,"
-        "    latitude numeric,"
+        "    longitude text,"
+        "    latitude text,"
         "    UNIQUE (stn_id, name)"
         ");"
         "CREATE TABLE datafiles ("
