### 2017-07-13 486bbc2
message when no files exist for given dates
