### 2017-07-13 2f51583
fix tests for downloading non-existent dates
