### 2017-07-14 #116 @mpadge
A general FYI: I had to upload a new version to CRAN because changes in the LA data were causing failures. I also addressed the first issue that @chucheria encountered so the package now gives the more informative message:
```
> dl_bikedata(city = 'la', dates = 201601:201605)
There are no la files for those dates
```
I was unable to repeat the second error, but note that `tempdir()` seems to have a double forward slash - `.../T//RtmpU6AJg6/` instead of presumably `.../T/RtmpU6AJg6/`. If that were the problem, I would have no idea of either ultimate cause or solution?
