### 2017-07-17 #116 @eamcvey
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is a nice package that made it easy to access the data as promised.  The linked article showing analysis of this type of data is helpful to make clear what can be done with it.  I was able to work through the vignette successfully and look at some additional things on my own.

Some specific issues I noticed:

- bike_db_totals and bike_demographic_data are lacking examples in the documentation
- There are spaces before parentheses for functions used in the vignette - I'm not sure what to make of this but it disturbs my sense of R code...
- This may be lack of understanding on my part, but vignette() could not find the package vignette locally
- I see the authors and contact info in the DESCRIPTION, but not in the README or CONTRIBUTING files - which sounds like a requirement per the reviewing guidelines
- When I first used this package store_bikedata worked fine, but more recently it throws Error: The dbplyr package is required to communicate with database backends.  I assume this is due to other package updates, but something to take into consideration.
