### 2017-07-19 #116 @mpadge
Thanks both @chucheria and @eamcvey for your reviews! :smiling_imp: I should have addressed all issues raised sometime next week, and will respond accordingly.
