### 2017-07-21 60f72a5
add conduct, contributing files
