+++ CONTRIBUTING.md
@@ -1,0 +1,55 @@
+# Contributing to bikedata
+
+## Opening issues
+
+The easiest way to note any behavioural curiosities or to request any new
+features is by opening a [github issue](https://github.com/mpadge/bikedata/issues).
+
+
+## Development guidelines
+
+If you'd like to contribute changes to `bikedata`, we use [the GitHub
+flow](https://guides.github.com/introduction/flow/index.html) for proposing,
+submitting, reviewing, and accepting changes. If you haven't done this before,
+there's a nice overview of git [here](http://r-pkgs.had.co.nz/git.html), as well
+as best practices for submitting pull requests
+[here](http://r-pkgs.had.co.nz/git.html#pr-make).
+
+The `bikedata` coding style diverges somewhat from [this commonly used R style
+guide](http://adv-r.had.co.nz/Style.html), primarily in the following two ways,
+both of which improve code readability: (1) All curly braces are vertically aligned:
+```r
+this <- function ()
+{
+    x <- 1
+}
+```
+and **not**
+```r
+this <- function(){
+    x <- 1
+}
+```
+and (2) Also highlighted in that code is the additional whitespace which
+permeates `bikedata` code. Words of text are separated by whitespace, and so
+code words should be too:
+```r
+this <- function1 (function2 (x))
+```
+and **not**
+```r
+this <- function1(function2(x))
+```
+with the natural result that one ends up writing
+```r
+this <- function ()
+```
+with a space between `function` and `()`. That's it.
+
+
+## Code of Conduct
+
+We want to encourage a warm, welcoming, and safe environment for contributing to
+this project. See the
+[code of conduct](https://github.com/mpadge/bikedata/blob/master/CONDUCT.md) for
+more information.
