### 2017-07-24 6fc46f3
update RcppExports filenames with underscore
