### 2017-07-24 bfcb368
extend example for bike_demographic_data
also fix demographic data as stored in database for bo, ch, which was previously wrong
