### 2017-07-24 #116 @mpadge
# Package changes in response to reviews

1. As already mentioned, the previously mis-leading message that `"All data files already exist"` has been replaced with `"There are no [city] files for that date."`
2. Repo now includes both `CONTRIBUTING.md` and `CONDUCT.md` - thanks to both reviewers for pointing out this absence on my part.
3. Both reviewers noted `dplyr` (or `dbplyr`) issues. These arose in response to the upgrade to `dplyr_0.7`, which led to interesting situations for many package developers.  `bikedata` has resolved all such issues by dropping the `dplyr` dependency.
4. Extensive example now provided for `bike_db_totals()` - thanks! An example was previously included in `bike_demographic_data()`, but just one single line that was presunably easy to overlook. The reviewer's comment nevertheless inspired me to provide a much more extensive example, during which I also relealised that there were some formatting issues for demographic data from some cities as stored in the database. These have all been rectified.

# Changes not yet implemented

1. I was unable to reproduce the problem encountered by @chucheria, but have already noted that I have previously encountered strange behaviour when running `store_bikedata()` multiple times. I suspect that the main cause of such problems is index creation, which may only be performed once for a database, and any attempt to create indices that already exist will throw an error. Current
[`bikedata` issue#38](https://github.com/mpadge/bikedata/issues/38) should largely resolve these issues.

# Responses to other comments not related to package changes

1. There was always `NEWS.md` file which I presume the first reviewer simply overlooked.
2. The package uses `GPL-3` which is a copyright license and therefore requires no `LICENSE` file. (LICENSE files are also rejected by CRAN for GPL-3 packages: they only apply to non-copyright licenses such as MIT).
3. Yeah, i know file names are hyphenated rather than underscored, but this is done because there are systems which always interpret underscores as non-alpha-numeric characters and therefore produce inconsistent highlighting of function names; underscores avoid this problem. (That comment only applies to those who use `vim`, `emacs`, or other non-`RStudio` interfaces.)
4. Yes, I do use `function (x)` rather than `function(x)`, but now explain this in `CONTRIBUTING.md`: "Words of text are separated by whitespace, and so code words should be too." Let's just say I'm pre-empting physchological analyses of R-coding style akin to what has been done for R colour palettes. Spaces will one day reign ...
5. Regarding authorship, the ropensci authorship guide merely describes the requirements for `DESCRIPTION`, to which this package accords, so I think current status is okay in that regard. I'll happily modify if required or desired.
6. Finally, note that `vignette()` will *not* automatically load the vignette when the package is psuedo-loaded with `devtools::load_all()`. I nevertheless checked with a "proper" locall load (`R CMD install <tarball>`) and was then able to open `vignette()` as expected.

Thanks again @eamcvey and @chucheria for helpful insight into improving the package!
