### 2017-07-24 #116 @mpadge
Thanks @maelle! That change is [issue#38](https://github.com/mpadge/bikedata/issues/38), which I will definitely resolve within the coming week. I'll let you know as soon as it's done. :bike: :biking_man: :biking_woman: :bike:
