### 2017-07-25 #116 @maelle
@mpadge For info you can already add the rOpenSci badge to the README of `bikedata`, for now it'll show "under review2" and after approval "peer-reviewed".
