### 2017-07-25 2690c3d
add ropensci badge
