+++ R/database-stats.R
@@ -460,6 +460,7 @@ bike_daily_trips <- function (bikedb, city, station, member, birth_year, gender,
 #' @examples
 #' bike_demographic_data ()
 #' # Examples of filtering data by demographic parameters:
+#' /dontrun{
 #' data_dir <- tempdir ()
 #' bike_write_test_data (data_dir = data_dir)
 #' bikedb <- file.path (data_dir, 'testdb')
@@ -479,6 +480,7 @@ bike_daily_trips <- function (bikedb, city, station, member, birth_year, gender,
 #' sum (bike_tripmat (bikedb = bikedb, city = 'dc', birth_year = 1990))
 #' bike_rm_test_data (data_dir = data_dir)
 #' bike_rm_db (bikedb)
+#' }
 bike_demographic_data <- function ()
 {
     dat <- data.frame (city = c ('bo', 'ch', 'dc', 'la', 'lo', 'ny', 'ph'),
