+++ man/bike_demographic_data.Rd
@@ -16,7 +16,7 @@ Static summary of which systems provide demographic data
 \examples{
 bike_demographic_data ()
 # Examples of filtering data by demographic parameters:
-/dontrun{
+\dontrun{
 data_dir <- tempdir ()
 bike_write_test_data (data_dir = data_dir)
 bikedb <- file.path (data_dir, 'testdb')
