### 2017-07-28 #116 @chucheria
Thanks for your comments @mpadge. I'm sorry I overlooked the `NEWS` file! So embarrasing... I didn't know you don't need a license, anyway if CRAN doesn't admit it better this way, I guess I'm used to do the `LICENSE` and `gitignore` first thing.

About the error you commented you couldn't reproduce, I'll try to do the same thing as soon as I have some time in case it appears again an open an issue with more info in case it appears, does this sound good?

Thank you both @mpadge and @maelle! This has been fun!
