### 2017-07-30 5cd8654
add separate index_bikedata_db fn
