+++ NAMESPACE
@@ -12,6 +12,7 @@ export(bike_summary_stats)
 export(bike_tripmat)
 export(bike_write_test_data)
 export(dl_bikedata)
+export(index_bikedata_db)
 export(store_bikedata)
 importFrom(RSQLite,SQLite)
 importFrom(RSQLite,dbBind)
