### 2017-07-31 cc31994
no more auto-indexing on store_bikedata
use index_bikedata_db instead
