### 2017-07-31 mpadge/bikedata#38 @mpadge
mostly done with [this commit](https://github.com/mpadge/bikedata/commit/cc31994e928f689779d23d6eaf1e09f048ba4f9d) but issue kept open for potential subsequent implementation of demographic indexes.
