### 2017-08-03 #116 @chucheria
No error, the package seems to be solid.
