### 2017-08-10 #116 @maelle
@mpadge sorry for the delay! The package is now approved, great work, and thanks @eamcvey + @chucheria for your reviews!

To-dos:
- [ ] Transfer the repo to the rOpenSci organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.
- [ ] OPTIONAL You can add the reviewers in DESCRIPTION as "rev" if you want to (and if they agree). See e.g. [this example](https://github.com/ropensci/stplanr/blob/master/DESCRIPTION). If you do this you'll need to build the package on R-devel for CRAN submission (because "rev" is now accepted in R-devel)

- [ ] Add the (brand new!) "Peer-reviewed" badge to your README (in the future people will add this on submittal; it updates through the stages of review):
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (we'll turn on the services on our end.)
- [ ] Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). Let me know if you are interested.
