### 2017-08-12 ropensci/bikedata#45 @tombuckley
link in github project header should be to ropensci.github.. rather than mpadge.github...
should be: https://ropensci.github.io/bikedata/
