### 2017-08-13 07dd057
github handover mpadge -> ropensci
closes #45 - thanks @tombuckley for being impressively on the ball!
