+++ DESCRIPTION
@@ -36,8 +36,8 @@ LinkingTo:
 VignetteBuilder: knitr
 SystemRequirements: C++11
 NeedsCompilation: yes
-URL: https://github.com/mpadge/bikedata
-BugReports: https://github.com/mpadge/bikedata/issues
+URL: https://github.com/ropensci/bikedata
+BugReports: https://github.com/ropensci/bikedata/issues
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
