+++ _pkgdown.yml
@@ -50,4 +50,4 @@ navbar:
     href: reference/index.html
   right:
   - icon: fa-github fa-lg
-    href: https://github.com/mpadge/bikedata
+    href: https://github.com/ropensci/bikedata
