+++ R/database-stats.R
@@ -61,10 +61,15 @@ bike_cities_in_db <- function (bikedb)
 #' @noRd
 get_new_datafiles <- function (bikedb, flist_zip)
 {
-    db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
-    old_files <- RSQLite::dbReadTable (db, 'datafiles')$name
-    RSQLite::dbDisconnect (db)
-    flist_zip [which (!basename (flist_zip) %in% old_files)]
+    if (!is.null (flist_zip))
+    {
+        db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
+        old_files <- RSQLite::dbReadTable (db, 'datafiles')$name
+        RSQLite::dbDisconnect (db)
+        flist_zip [which (!basename (flist_zip) %in% old_files)]
+    }
+
+    return (flist_zip)
 }
 
 #' Get dates of first and last trips, and number of days in between, for each
