### 2017-08-14 8bf21d0
RSQLite -> DBI; closes #43
