+++ R/database-stats.R
@@ -15,9 +15,9 @@
 #' @noRd
 indexes_exist <- function (bikedb)
 {
-    db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
-    idx_list <- RSQLite::dbGetQuery (db, "PRAGMA index_list (trips)")
-    RSQLite::dbDisconnect (db)
+    db <- DBI::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
+    idx_list <- DBI::dbGetQuery (db, "PRAGMA index_list (trips)")
+    DBI::dbDisconnect (db)
     nrow (idx_list) > 2 # 2 because city index is automatically created
 }
 
@@ -28,9 +28,9 @@ indexes_exist <- function (bikedb)
 #' @noRd
 num_datafiles_in_db <- function (bikedb)
 {
-    db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
-    numtrips <- RSQLite::dbGetQuery (db, "SELECT Count(*) FROM datafiles")
-    RSQLite::dbDisconnect (db)
+    db <- DBI::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
+    numtrips <- DBI::dbGetQuery (db, "SELECT Count(*) FROM datafiles")
+    DBI::dbDisconnect (db)
     return (as.numeric (numtrips))
 }
 
@@ -41,9 +41,9 @@ num_datafiles_in_db <- function (bikedb)
 #' @noRd
 bike_cities_in_db <- function (bikedb)
 {
-    db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
-    cities <- RSQLite::dbGetQuery (db, "SELECT city FROM stations")
-    RSQLite::dbDisconnect (db)
+    db <- DBI::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
+    cities <- DBI::dbGetQuery (db, "SELECT city FROM stations")
+    DBI::dbDisconnect (db)
     cities <- unique (cities)
     rownames (table (cities)) # TODO: Find a better way to do that
 }
@@ -63,9 +63,9 @@ get_new_datafiles <- function (bikedb, flist_zip)
 {
     if (!is.null (flist_zip))
     {
-        db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
-        old_files <- RSQLite::dbReadTable (db, 'datafiles')$name
-        RSQLite::dbDisconnect (db)
+        db <- DBI::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
+        old_files <- DBI::dbReadTable (db, 'datafiles')$name
+        DBI::dbDisconnect (db)
         flist_zip [which (!basename (flist_zip) %in% old_files)]
     }
 
@@ -84,13 +84,13 @@ get_new_datafiles <- function (bikedb, flist_zip)
 #' @noRd
 bike_station_dates <- function (bikedb, city)
 {
-    db <- RSQLite::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
+    db <- DBI::dbConnect (RSQLite::SQLite(), bikedb, create = FALSE)
     qry <- paste0 ("SELECT MIN (STRFTIME('%Y-%m-%d', start_time)) AS 'first',",
                    "MAX (STRFTIME('%Y-%m-%d', start_time)) AS 'last',",
                    "start_station_id AS 'station' FROM trips WHERE city = '",
                    city, "' GROUP BY start_station_id")
-    dates <- RSQLite::dbGetQuery (db, qry)
-    RSQLite::dbDisconnect (db)
+    dates <- DBI::dbGetQuery (db, qry)
[…]
