+++ DESCRIPTION
@@ -15,6 +15,7 @@ License: GPL-3
 Depends:
     R (>= 3.0)
 Imports:
+    DBI,
     httr,
     lubridate,
     magrittr,

[…]
