### 2017-08-14 b500316
version bump -> 0.0.4
