### 2017-08-15 a02e07d
add ropensci revs to DESCR
