### 2017-08-15 92036f1
fix DESCR
