### 2017-08-17 c0a7422
update NEWS
