### 2017-08-17 #116 @mpadge
Thanks @maelle for all of the help and guidance getting the package through to this exciting last stage!
You list of TODOs:
- [x] Repo transferred
- [x] both @chucheria and @eamcvey added to DESCRIPTION of new release (0.0.4) with role = "rev"
- [x] ropensci badge added
- [x] all other badges re-directed (see [#45](https://github.com/ropensci/bikedata/issues/45))
- [x] Zenodo now watching repo
- [x] First release on zenodo [here](https://zenodo.org/account/settings/github/repository/ropensci/bikedata) with [doi:10.5281/zenodo.844466](https://doi.org/10.5281/zenodo.844466)
- [ ] Submit to JOSS - I thought that happened automatically? (It did for `osmdata`) Let me know if i need to do anything there

Finally, re: blog post: absolutely interested! Yes please!
