### 2017-08-17 #116 @maelle
You mean you didn't have to submit it? :thinking: Having @karthik as an editor of the R package surely helped. :wink:

In my experience you need to submit it, choose Karthik as JOSS editor.

Reg. the blog post let me tag @stefaniebutland who'll get in touch with you.
