### 2017-08-17 #116 @mpadge
okay, i've submitted it now, and put Karthik as the editor:
- [x] Submit to JOSS

ping @richardellison, so you know where we're at here. new publication pending!
