### 2017-08-20 #116 @mpadge
YIPPEE!!! 🚲 🚲 🚲 I didn't realise until I just discovered the [ropensci issue on this](https://github.com/ropensci/onboarding/issues/125) that i should have but didn't actually request the explicit approval of @chucheria and @eamcvey for me to insert those `rev` roles in the [DESCRIPTION](https://github.com/ropensci/bikedata/blob/master/DESCRIPTION). May I kindly request retrospective approval now? Please!
