### 2017-08-23 #116 @stefaniebutland
@mpadge (I'm responding later than I had hoped)
Great to hear you're interested in contributing a blog post about this package. I hope it gets more eyes on it and generates some enthusiasm and input.

For rOpenSci blog post you could do either a short post for the [Developer Blog](https://ropensci.org/tech-notes/) that has a technical focus, or do a post for the [main blog](https://ropensci.org/blog/) whose audience is broader.

Main blog post would include more narrative about your motivation for creating the package, unmet need, how-to use, good to end with a thank you to package reviewers with links to their GitHub or Twitter, point readers to issues and what you think is next to improve the package and invite people to open or address an issue etc.

Deadlines:
- Developer blog post can come any time.
- For main blog we post about once a week. Next open slots are late September. We ask for a draft submitted as a pull request a week in advance.

Practical instructions:
- https://github.com/ropensci/roweb/blob/master/.github/CONTRIBUTING.md
- suggested tags: R, community, software, review, onboarding, package, package_name, topic labels used in onboarding review

Which type of post are you thinking of?

I had been thinking of inviting you to contribute a post about `osmdata`, so if you're interested you could do both as developer blog style, or one dev and one main.

Did I miss anything?
