### 2017-08-23 #116 @mpadge
thanks @stefaniebutland, and no worries about delayed response. I'm absolutely keen to get these packages known in any and all ways possible. Absolutely happy for at least one to be a main blog entry, for which my preference would be `bikedata`, with `osmdata` then shucked off to a dev blog post. If that's okay with you, i'll happily get cracking on a PR for main blog post on bikedata
